pyproject.toml
[tool.poetry]
name = "tgparser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
telethon = "^1.40.0"
asyncio = "^3.4.3"
nest-asyncio = "^1.6.0"
pymongo = "^4.13.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
motor = "^3.7.1"
asyncpg = "^0.30.0"
aiohttp = "^3.12.14"
aiogram = "^3.21.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
